requirements.txt
akismet==1.0.1
certifi==2017.7.27.1
chardet==3.0.4
Django==2.2.12
django-appconf==1.0.2
django-braces==1.11.0
django-ckeditor==5.9.0
django-contact-form==1.4.1
django-geojson==2.10.0
django-imagekit==4.0.1
django-js-asset==1.2.2
django-leaflet==0.22.0
django-maintenancemode==0.11.2
django-model-utils==3.0.0
django-ranged-response==0.2.0
django-registration==2.2
django-registration-redux==2.7
django-simple-captcha==0.5.12
django-star-ratings==0.9.1
idna==2.5
IPy==0.83
jsonfield==2.0.1
model-mommy==1.3.2
olefile==0.44
pilkit==2.0
Pillow==7.1.1
python-magic==0.4.13
pytz==2017.2
PyYAML==5.3.1
requests==2.23.0
six==1.10.0
sqlparse==0.3.1
swapper==1.1.2.post1
twitter==1.17.1
urllib3==1.25.9
